correct some logic errors

+++ python/cugraph/cugraph/structure/graph_implementation/simpleDistributedGraph.py
@@ -83,7 +83,9 @@ class simpleDistributedGraphImpl:
             elif values.dtype == 'int64':
                 values = values.astype('float64')
         else:
-            values = cudf.Series(cupy.ones(len(edata_x[0])))
+            values = cudf.Series(
+                cupy.ones(len(edata_x[0]), dtype='float32')
+            )
         
         if simpleDistributedGraphImpl.edgeIdCol in edata_x[0]:
             if simpleDistributedGraphImpl.edgeTypeCol not in edata_x[0]:
